Initial restructuring of architecture
