+++ assets/js/main.js
@@ -2,3 +2,11 @@
 This file is for linking and responsiveness purposes only. 
 NO CHESS LOGIC HERE OTHER THAN INITIALIZATION CALLS.
 */
+
+
+
+ 
+  
+
+  
+
